% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdownbug.R
\name{ComparisonMethods}
\alias{ComparisonMethods}
\alias{<,MyType,numeric-method}
\alias{<=,MyType,numeric-method}
\alias{>,MyType,numeric-method}
\alias{>=,MyType,numeric-method}
\title{Comparison methods}
\usage{
\S4method{<}{MyType,numeric}(e1, e2)

\S4method{<=}{MyType,numeric}(e1, e2)

\S4method{>}{MyType,numeric}(e1, e2)

\S4method{>=}{MyType,numeric}(e1, e2)
}
\description{
Generic methods and built-in functions for MyType objects
}
\section{Functions}{
\itemize{
\item \code{e1 < e2}: Less than comparison

\item \code{e1 <= e2}: Less than comparison

\item \code{e1 > e2}: Less than comparison

\item \code{e1 >= e2}: Less than comparison

}}
